+++ src/main/java/printbot/commands/AddCommand.java
@@ -1,0 +1,18 @@
+package printbot.commands;
+
+import printbot.tasks.Task;
+import printbot.tasks.TaskList;
+import printbot.ui.UI;
+
+public class AddCommand extends Command {
+
+    public void execute(UI ui, TaskList taskList, Task task) {
+        taskList.addTask(task);
+        ui.uiAddTask(task, taskList);
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
