+++ src/main/java/printbot/commands/Command.java
@@ -1,0 +1,12 @@
+package printbot.commands;
+
+import printbot.tasks.TaskList;
+import printbot.ui.UI;
+
+public abstract class Command {
+
+    //all must have execute(...);
+
+    public abstract boolean isExit();
+
+}
