+++ src/main/java/printbot/exceptions/CorruptedSaveException.java
@@ -1,4 +1,4 @@
-package PrintBot;
+package printbot.exceptions;
 
 // Level-7
