+++ src/main/java/printbot/exceptions/PrintException.java
@@ -1,4 +1,4 @@
-package PrintBot;
+package printbot.exceptions;
 
 // Level 5
