+++ src/main/java/printbot/storage/Storage.java
@@ -1,5 +1,9 @@
-package PrintBot;
+package printbot.storage;
 
+
+import printbot.exceptions.CorruptedSaveException;
+import printbot.tasks.Task;
+import printbot.tasks.TaskList;
 
 import java.io.*;
 import java.util.List;
