+++ src/main/java/printbot/tasks/TaskList.java
@@ -1,8 +1,7 @@
-package PrintBot;
+package printbot.tasks;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.io.*;
 
 public class TaskList {
