+++ src/main/java/printbot/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package PrintBot;
+package printbot.tasks;
 
 /*
  * Store task name
